Added Functions.h and .cpp files
